Harmonized with dbAbstraction
